Move MediaDownloadTask to separate file
